xoa booking user

+++ ABnB/src/main/java/com/TDD/ABnB/controller/AppBookingController.java
@@ -3,8 +3,10 @@ package com.TDD.ABnB.controller;
 import com.TDD.ABnB.exceptions.DuplilcateUserException;
 import com.TDD.ABnB.models.AppBooking;
 import com.TDD.ABnB.models.AppProperty;
+import com.TDD.ABnB.models.AppUser;
 import com.TDD.ABnB.services.app_booking_service.AppBookingService;
 import com.TDD.ABnB.services.app_property_service.AppPropertyService;
+import com.TDD.ABnB.services.app_user_service.AppUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,6 +14,9 @@ import org.springframework.web.bind.annotation.*;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.time.Period;
+import java.time.format.DateTimeFormatter;
 import java.util.Date;
 
 @RestController
@@ -21,8 +26,12 @@ public class AppBookingController {
 
     @Autowired
     private AppBookingService appBookingService;
+
     @Autowired
     private AppPropertyService appPropertyServiceImpl;
+
+    @Autowired
+    private AppUserService appUserService;
 
     @GetMapping()
     public ResponseEntity<Iterable<AppBooking>> showListBooking() {
@@ -52,14 +61,20 @@ public class AppBookingController {
         Date checkOutDate = new SimpleDateFormat("yyyy/MM/dd").parse(appBooking.getCheckoutDate());
         Iterable<AppBooking> appBookings = appBookingService.findAllByAppProperty(appPropertyServiceImpl.findById(appBooking.getAppProperty().getId()));
         for (AppBooking booking : appBookings) {
-            Date checkOutDateEx = new SimpleDateFormat("yyyy/MM/dd").parse(booking.getCheckinDate());
+            Date checkOutDateEx = new SimpleDateFormat("yyyy/MM/dd").parse(booking.getCheckoutDate());
+            Date checkInDateEx = new SimpleDateFormat("yyyy/MM/dd").parse(booking.getCheckinDate());
             boolean bookingDateValid = checkInDate.after(checkOutDate);
-            boolean checkinDateValid = checkInDate.after(checkOutDateEx);
+            boolean checkInDateValid = checkInDate.after(checkOutDateEx);
+            boolean checkOutDateValid = checkInDate.after(checkInDateEx);
 
-            if (!bookingDateValid && !checkinDateValid) {
+            if ((!bookingDateValid) && ((!checkInDateValid) || (!checkOutDateValid))) {
                 throw new DuplilcateUserException("Booking unavailable");
             }
         }
+
+        System.out.println(appUserService.findById(appBooking.getAppUser().getId()).getAppBookings());
+        System.out.println(appPropertyServiceImpl.findById(appBooking.getAppProperty().getId()).getAppBookings());
+
         appBookingService.save(appBooking);
         ResponseEntity<AppBooking> res = new ResponseEntity<>(appBooking, HttpStatus.ACCEPTED);
         return res;
@@ -73,8 +88,18 @@ public class AppBookingController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteBooking(@PathVariable("id") Long id) {
+    public ResponseEntity<AppUser> deleteBooking(@PathVariable("id") Long id) throws DuplilcateUserException {
         AppBooking appBooking = appBookingService.findById(id);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MM/dd");
+        LocalDate localDate = LocalDate.parse(appBooking.getCheckinDate(), formatter);
+        Period period = Period.between(localDate, java.time.LocalDate.now());
+        System.out.println(period.getDays());
+        if (period.getDays() <= 1) {
+            throw new DuplilcateUserException("Cannot cancel booking this late");
+        }
+        AppUser appUser = appUserService.findById(appBooking.getAppUser().getId());
         appBookingService.delete(appBooking);
+        ResponseEntity<AppUser> res = new ResponseEntity<>(appUser, HttpStatus.ACCEPTED);
+        return res;
     }
 }
